+++ pvi/utils/__init__.py
@@ -1,3 +1,6 @@
-from .BrFile import BrFile
+from .BrFile import BrFile, ModuleType
+from .BrDataObjectFile import BrDataObjectFile
+from .BrMotionDataObjectFile import BrMotionDataObjectFile
+from .BrXMLBasedConfigFile import BrXMLBasedConfigFile
 from .BrLoggerFile import BrLoggerFile
 from .BrProfilerDataFile import BrProfilerDataFile